requirements/base.txt
# Web Framework
fastapi==0.121.1
uvicorn[standard]==0.38.0
pydantic==2.12.0

# Database
sqlmodel==0.0.14
sqlalchemy==2.0.44
alembic==1.17.1
psycopg2-binary==2.9.11

# Authentication
python-jose[cryptography]==3.5.0
passlib[argon2]==1.7.4
argon2-cffi==25.1.0
python-multipart==0.0.6

# Configuration
pydantic-settings==2.0.3
python-dotenv==1.0.0


# Rate Limiting
slowapi==0.1.9


# Media Processing
Pillow==11.3.0
python-magic==0.4.27
aiofiles==23.2.1
ffmpeg-python==0.2.0

# HTTP Client
httpx==0.28.1

# OIDC / OAuth
authlib==1.6.5
redis==5.0.1
itsdangerous==2.1.2  # Required by Authlib for session state management

# Task Queue
celery==5.4.0

# Logging - using standard Python logging

# Validation
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
psutil==5.9.6
